refactor: 앨범 상세 ViewModel 분리 #149
